Implemented Basic API Mock Up
Added serde and serde_json crates.

Created findUserById(i64) in database.rs.

Created api module with structs.rs to hold data transport structs for API communication. These will be serialized as JSON strings while in transit.

Added /login route to main.rs utilizing the above changes to check if the provided UserId exists. If so, return the UserId and a greeting message. Otherwise, do not return a UserId and provide an error message.