Iterating ManageUsers UI
Added wasm-logger and log to help with developing the Dioxus web app.
Updated src/routes/admin.rs#userList to accept an args parameter, which will contain the URL encoded POST argument list.
Iterated app.rs to properly request the list of users, parse the response, and display the list of users in the rsx.