Pivoting Sea Orm Usage
While it was fun exploring functions to use sea-orm Entities in generic ways, my eyes opened to how little it was really helping in terms of code that needed to be typed. So, removed most methods out of database.rs. Now it houses getDatabaseConnection() and I moved createTestDatabase() into the internal tests module. Set database::tests to be public.
Refactored the uses of the old database functions into normal sea-orm calls.